edit some style and remake loader
