+++ src/scripts/index.js
@@ -2,7 +2,6 @@
 /* eslint-disable import/no-named-as-default-member */
 import 'regenerator-runtime'; /* for async await transpile */
 import '../styles/main.css';
-import '../styles/aside.css';
 import '../styles/responsive.css';
 import 'bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
