+++ src/scripts/views/component/custom-loader.js
@@ -6,9 +6,24 @@ class Loader extends HTMLElement {
 
   render() {
     this.innerHTML = `
-        <div class="loader"></div>
+    <div class="dots"></div>
           `;
   }
 }
 
 customElements.define('custom-loader', Loader);
+
+// /* eslint-disable linebreak-style */
+// class Loader extends HTMLElement {
+//   connectedCallback() {
+//     this.render();
+//   }
+
+//   render() {
+//     this.innerHTML = `
+//         <div class="loader"></div>
+//           `;
+//   }
+// }
+
+// customElements.define('custom-loader', Loader);
