+++ src/styles/admin-dashboard.css
@@ -26,7 +26,7 @@
 
 .card__item {
   display: flex;
-  justify-content: space-around;
+  justify-content: space-between;
   padding: 30px;
   border-radius: 10px;
   color: #fff;
