+++ src/styles/responsive.css
@@ -16,6 +16,7 @@
     transform: translate(-100%, 0);
     
   }
+
   .item-news__content p {
     font-size: 0.9rem;
     
@@ -142,13 +143,27 @@
     opacity: 0;
     right: -33px;
     min-width: 120px;
-    background-color: #005555;
+    background-color: white;
     transition: all 0.7s ease-in-out;
     box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
     transition: visibility 0s, opacity 0.3s linear;
     z-index: 4;
     
     border-radius: 0 0 5px 5px;
+  }
+    
+  .hero__title {
+    font-size: 2.9em;
+    font-weight: lighter;
+  }
+  .hero__inner{
+    flex-direction: row !important;
+  }
+  .hero__inner__text {
+   align-items: flex-start;
+  }
+  .hero__title {
+    text-align: left;
   }
   .btn-drawer {
     display: none;
@@ -299,34 +314,12 @@
   }
 }
 
-@media screen and (min-width: 600px) {
-  .hero__inner{
-    padding-left: 380px !important;
-  }
-}
 
 @media screen and (min-width: 800px) {
-  .hero__inner{
-    padding-left: 490px !important;
-  }
 
   .hero__highlight {
     font-size: 50px !important;
   }
   
-  .hero__title {
-    font-size: 50px !important;
-  }
-}
 
-@media screen and (min-width: 900px) {
-  .hero__inner{
-    padding-left: 580px !important;
-  }
-}
-
-@media screen and (min-width: 1100px) {
-  .hero__inner{
-    padding-left: 800px !important;
-  }
 }
